Moved cr.webservices functions into cr.utils.web module.
New stuff this afternoon about web services and the online part, stay tuned!
